fix: custom.test.ts を廃止して個別にテストファイルを用意して実装する形で実装
