+++ packages/zenn-markdown-html/__tests__/custom-syntax/embed/blueprintue.test.ts
@@ -1,18 +1,16 @@
 import { parse } from 'node-html-parser';
-import markdownToHtml from '../../src/index';
+import markdownToHtml from '../../../src/index';
 
-describe('CodeSandBox埋め込み要素のテスト', () => {
-  const validUrl =
-    'https://codesandbox.io/embed/test-example?fontsize=14&hidenavigation=1&theme=dark';
-  const invalidUrl =
-    'https://bad-example.codesandbox.io/embed/test-example?fontsize=14&hidenavigation=1&theme=dark';
+describe('Blueprintue埋め込み要素のテスト', () => {
+  const validUrl = 'https://blueprintue.com/render/examples';
+  const invalidUrl = 'https://bad-example.blueprintue.com/render/examples';
 
   describe('デフォルトの挙動', () => {
     describe('有効なURLの場合', () => {
       test('<iframe />に変換する', () => {
-        const html = markdownToHtml(`@[codesandbox](${validUrl})`);
+        const html = markdownToHtml(`@[blueprintue](${validUrl})`);
         const iframe = parse(html).querySelector(
-          `span.embed-codesandbox iframe`
+          `span.embed-blueprintue iframe`
         );
 
         expect(iframe?.attributes).toEqual(
@@ -23,21 +21,21 @@ describe('CodeSandBox埋め込み要素のテスト', () => {
 
     describe('無効なURLの場合', () => {
       test('エラーメッセージを出力する', () => {
-        const html = markdownToHtml(`@[codesandbox](${invalidUrl})`);
+        const html = markdownToHtml(`@[blueprintue](${invalidUrl})`);
 
         expect(html).toContain(
-          '「https://codesandbox.io/embed/」から始まる正しいURLを入力してください'
+          '「https://blueprintue.com/render/」から始まる正しいURLを指定してください'
         );
       });
     });
   });
 
-  describe('customEmbed.codesandbox()を設定している場合', () => {
+  describe('customEmbed.blueprintue()を設定している場合', () => {
     test('渡した関数を実行する', () => {
       const customizeText = 'customized text!';
       const mock = jest.fn().mockReturnValue(customizeText);
-      const html = markdownToHtml(`@[codesandbox](${validUrl})`, {
-        customEmbed: { codesandbox: mock },
+      const html = markdownToHtml(`@[blueprintue](${validUrl})`, {
+        customEmbed: { blueprintue: mock },
       });
 
       expect(mock).toBeCalled();
