+++ packages/zenn-markdown-html/__tests__/custom-syntax/embed/speakerdeck.test.ts
@@ -1,5 +1,5 @@
 import { parse } from 'node-html-parser';
-import markdownToHtml from '../../src/index';
+import markdownToHtml from '../../../src/index';
 
 describe('SpeakerDeck埋め込み要素のテスト', () => {
   const validToken = 'example-token';
